Смещение в потоке 64-битное: nonce не заворачивается, и соединение не рвётся ради этого
Вопрос «как понять, что рукопожатия повторяются» вскрыл настоящую беду, а не пробел в журнале.
В журнале тишина потому, что в режиме потока смена ключей не происходила почти никогда: клиент
менял их после гигабайта отправки, а ХАБ не менял вообще — ветка потока в sendTo возвращалась
раньше, чем управление доходило до RetireDue. Между тем nonce в потоке — это смещение байта, и
оно было 32-битным: на четвёртом гигабайте счётчик заворачивается и тот же nonce уходит с тем же
ключом. Для AES-GCM и ChaCha20-Poly1305 это не ослабление, а полная потеря стойкости —
вскрывается и подделка, и содержимое. Четыре гигабайта на скачивании — один фильм, то есть
достижимо, и в направлении хаб→клиент защиты не было никакой.

Первым решением был ретайр: закрывать поток на подходе к границе. Оно неверно по цене — живой
туннель падает каждый гигабайт, внутренние сессии TCP рвутся на ровном месте. Правильный ответ
взят у veil: там порядковый номер записи восьмибайтный (record/v1: SeqSize = 8), и вопроса
заворота не возникает вовсе, а ротация ключей эпохами (epoch.Advance ратчетит корень, номер
эпохи входит в контекст KDF) служит прямой секретности и идёт без нового рукопожатия.

Здесь достаточно первой половины: смещения txOff и rxOff стали 64-битными. На проводе НЕ
меняется ничего — смещение не передаётся, обе стороны считают байты сами, а Seal и Open
принимали nonce как uint64 с самого начала; для первых четырёх гигабайт значения те же, что и
были. Поэтому ретайр по объёму в режиме потока убран с обеих сторон: рвать соединение больше не
за чем.

Проверено на живом хабе: 5 ГиБ через ОДНО соединение, границу 4 ГиБ прошли на 31-й секунде,
1072 Мбит/с, ни одного разрыва и ни одного повторного рукопожатия (их так и осталось четыре, по
одному на соединение), в журналах обеих сторон чисто. Стенд закрепляет само свойство:
TestStreamOffsetDoesNotWrap проверяет, что счётчик перешагивает 2^32 и продолжает расти.

Ветка поддельного TCP ретайрится по-прежнему, и там это неизбежно: nonce берётся из
относительного номера последовательности TCP, а он 32-битный по своей природе.

Периодическая смена ключей ради прямой секретности остаётся отдельной задачей, и делать её надо
эпохами, как в veil, — без разрыва соединения.

+++ hub/worker.go
@@ -561,12 +561,16 @@ func (w *worker) sendTo(d *session, row []byte, plen int) {
 	}
 	// Сессия по НАСТОЯЩЕМУ потоку: резать запись на сегменты не нужно, это работа ядра.
 	if d.st != nil {
+		// Заворота nonce здесь больше не может быть: смещение в потоке 64-битное (см. шапку
+		// wire.Stream), поэтому ни закрывать поток, ни менять ключи по объёму не нужно. Ветка
+		// поддельного TCP ниже ретайрится по-прежнему, и там это неизбежно: nonce берётся из
+		// относительного номера последовательности TCP, а он 32-битный по своей природе.
 		rec := row[wire.HdrRoom-wire.RecHdr : wire.HdrRoom]
-		err := d.st.WriteRecord(row, wire.RecHdr+plen+wire.Tag, func(rel uint32) error {
+		err := d.st.WriteRecord(row, wire.RecHdr+plen+wire.Tag, func(rel uint64) error {
 			if err := wire.RecBuild(rec, plen+wire.Tag); err != nil {
 				return err
 			}
-			_, e := d.tx.Seal(row[wire.HdrRoom:wire.HdrRoom+plen+wire.Tag], plen, rec, uint64(rel))
+			_, e := d.tx.Seal(row[wire.HdrRoom:wire.HdrRoom+plen+wire.Tag], plen, rec, rel)
 			return e
 		})
 		if err != nil {
